Removed common stuff into feed.js file
